requirements.txt
# FastAPI 核心依賴
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 認證相關
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI 模型整合
httpx==0.25.2

# 配置管理
python-dotenv==1.0.0
pydantic-settings==2.1.0

# 測試相關
pytest==7.4.3
pytest-asyncio==0.21.1

# 開發工具
black==23.11.0
isort==5.12.0
flake8==6.1.0

# 前端資源
jinja2==3.1.2
aiofiles==23.2.1
